:truck: [#326] -- move camunda options serializer into separate module
